+++ include/utils/Misc.hpp
@@ -8,10 +8,10 @@ private:
 
 public:
 	/* functions */
-	static double logBaseX (double base, double value);
-	static bool outOfBounds (int x, int y);
-	static int charToInt (char c);
-	static String numToWord (int num);
+	static float logBaseX(float base, float value);
+	static bool outOfBounds(int x, int y);
+	static int charToInt(char c);
+	static String numToWord(int num);
 };
 
 #endif
